pyproject.toml
[project]
name = "to-do-list-project-phase-1"
version = "0.1.0"
description = "A CLI ToDo list app with Python OOP, PostgreSQL via SQLAlchemy, Alembic migrations, and scheduled background commands"
authors = [
    {name = "Mehdi Gholami"}
]
readme = "README.md"
requires-python = "^3.10"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "sqlalchemy (>=2.0,<3.0)",
    "psycopg2-binary (>=2.9,<3.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "schedule (>=1.2.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
